requirements.txt
livekit-agents[deepgram,openai,cartesia,silero,turn-detector]~=1.0
livekit-plugins-noise-cancellation~=0.2
livekit-plugins-elevenlabs
python-dotenv
openai
faiss-cpu
numpy
langchain-openai
langchain-community
langchain-core